extmod/vfs_rom_file: Don't cast to ssize_t, it's already signed.

+++ extmod/vfs_rom_file.c
@@ -118,7 +118,7 @@ static mp_uint_t vfs_rom_file_ioctl(mp_obj_t o_in, mp_uint_t request, uintptr_t 
                 self->file_offset = self->file_size + s->offset;
             }
             if (self->file_offset > self->file_size) {
-                if ((ssize_t)s->offset < 0) {
+                if (s->offset < 0) {
                     // Seek to before the start of the file.
                     *errcode = MP_EINVAL;
                     return MP_STREAM_ERROR;
